Move aws and azure connect code is sub-modules
