+++ crates/core/tedge/src/cli/connect/mod.rs
@@ -2,6 +2,8 @@ pub use self::cli::*;
 pub use self::command::*;
 pub use self::error::*;
 
+mod aws;
+mod azure;
 mod c8y;
 mod cli;
 mod command;
